Show effects and no deprecation
In file 01_scikit.ipynb, depending on the specific data points generated with random noise, it may happen that the effect of overfitting is not visible and one would have to re-create the noisy data a few times to be able to see the effect. Therefore, a specific set of values which lead to overfitiing is suggested.

Also, for 04_text.ipynb, cross validation and grid search are now part of model_selection in scikit-learn. The old version gives deprecated warnings and threatens that they will even be removed soon.